[int64] Support for HP / Mana / End / Damage / Hate (#2091)
* Initial int64 work

* Hate 64 bit

* Update special_attacks.cpp

* Aggro / Damage / Hate int64

* NPC edit adjustments

* Fix bot compile

* More int64 adjustments

* More int64 references

* npcedit references

* aggrozone

* More int64 changes

* More int64 changes for damage

* Many more damage int64 references

* More spell damage int64 conversions

* HealDamage

* Damage fully working

* Remove debug

* Add migration

* More int64 adjustments

* Much wow, many int64

* More int64

* PR adjustments

+++ common/string_util.cpp
@@ -575,9 +575,9 @@ char *RemoveApostrophes(const char *s)
 	return NewString;
 }
 
-const char *ConvertArray(int input, char *returnchar)
+const char *ConvertArray(int64 input, char *returnchar)
 {
-	sprintf(returnchar, "%i", input);
+	sprintf(returnchar, "%lld", input);
 	return returnchar;
 }
