+++ zone/entity.h
@@ -460,7 +460,7 @@ public:
 	void    ClearWaterAggro(Mob* targ);
 	void	ClearFeignAggro(Mob* targ);
 	void	ClearZoneFeignAggro(Mob* targ);
-	void	AggroZone(Mob* who, uint32 hate = 0);
+	void	AggroZone(Mob* who, uint64 hate = 0);
 
 	bool	Fighting(Mob* targ);
 	void	RemoveFromHateLists(Mob* mob, bool settoone = false);
